fix: Fix logging out problem when frontends are changed
We CANNOT navigate two frontends with the same login.
The problem was that, when one of the frontends sent
the token, and the token was unverified, it removed that
token, which logged out both of the frontends.

So, the solution was: to allow both the frontends to
communicate with the backend with the same token (i.e.
maybe don't delete the authtoken when authorization error
happens) or use 2 different tokens for two different logins.

I chose the second approach for now.